motion correction and a bunch of assorted things

+++ CaImAn/motion_correct_only.py
@@ -41,17 +41,26 @@ from caiman.utils.utils import download_demo
 
 parser = argparse.ArgumentParser(description='Execute the CaImAn (Calcium Imaging Analysis) motion correction operation on a video recording (.tif file or .tiff stack).')
 parser.add_argument('infile', type=str, nargs=1, help='file name (for .tif file) or folder (for .tiff stack)')
-parser.add_argument('--outfile',type=str, nargs=1, default=[], help='file name under which to store the output hdf5 file (use .hdf5 suffix)')
+parser.add_argument('--outfile',type=str, nargs=1, default=[], help='file name under which to store the output motion correction movie')
+parser.add_argument('--mc_temp',type=str, nargs='?', default=None, help='template .tif file for motion correction')
 args = vars(parser.parse_args())
 
 infile = args['infile']
 outfile = args['outfile']
+mc_temp = args['mc_temp']
 
 print('Using arguments list: {}'.format(args))
 
 orig_in = infile[0]
 
-if os.path.isdir(infile[0]):
+
+if '.txt' in infile[0]:
+    with open(infile[0], 'r') as f:
+        flist = []
+        for line in f:
+            flist = flist + [line.replace('\n','')]
+    fname = flist
+elif os.path.isdir(infile[0]):
     tmpMovPath = os.path.join(infile[0], 'tmp_mov.tif')
     if not os.path.isfile(tmpMovPath):
         tfiles = glob.glob( os.path.join(infile[0], '*.tiff')) + glob.glob( os.path.join(infile[0], '*.tif'))
@@ -60,27 +69,41 @@ if os.path.isdir(infile[0]):
             for i in range(len(tfiles)):
                 if (i%100) == 0:
                     print("Writing movie to temp file, frame {}/{}".format(i,len(tfiles)))
-                im = tif.imread(tfiles[i])
+                im = tif.imread(tfiles[i]).astype('float16')
                 if len(im.shape)==2:
-                    writer.save(im, compress=6, photometric='minisblack')
+                    writer.save(im , contiguous=True, compress=6, photometric='minisblack')
                 else:
                     for j in range(len(im)):
-                        writer.save(im[j], compress=6, photometric='minisblack')
+                        writer.save(im[j], contiguous=True, compress=6, photometric='minisblack')
     fname = [tmpMovPath]
 else:
+    tmpMovPath = None
     fname = infile
 
-
-
-if len(outfile)==0:
-    if os.path.isdir(orig_in):
-        outfile = orig_in + '_mc'
-        in_file_ext = ''
+# set outfile name
+outfileStore = outfile
+outfile = []
+in_file_ext = []
+for orig_in in fname:
+    if len(outfileStore)==0:
+        if os.path.isdir(orig_in):
+            outfile = outfile + [orig_in + '_mc']
+            in_file_ext = in_file_ext + ['']
+        else:
+            in_filename, in_file_ext_new = os.path.splitext(orig_in)
+            in_file_ext = in_file_ext + [in_file_ext_new]
+            outfile = outfile + [in_filename + '_mc' + in_file_ext_new]
     else:
-        in_filename, in_file_ext = os.path.splitext(orig_in)
-        outfile = in_filename + '_mc' + in_file_ext
-else:
-    outfile = outfile[0]
+        outfile = outfileStore
+
+
+# load in template file
+if mc_temp is not None:
+    print('Loading motion correction reference template {}'.format(mc_temp),flush=True)
+    mc_temp = tif.imread(mc_temp)
+    if len(mc_temp.shape)>2:
+        mc_temp = mc_temp[0]
+
 
 #%%
 
@@ -92,9 +115,13 @@ def main():
 
     #%% First setup some parameters for data and motion correction
     
-    n_processes = 12
+    n_processes = 18
 
     # dataset dependent parameters
+
+    nChannels = 1
+    channelToKeep = 0
+
     fr = 30            # imaging rate in frames per second
     decay_time = 0.4    # length of a typical transient in seconds
     dxy = (2., 2.)      # spatial resolution in x and y in (um per pixel)
@@ -105,13 +132,19 @@ def main():
     # motion correction parameters
     pw_rigid = False       # flag to select rigid vs pw_rigid motion correction
     # maximum allowed rigid shift in pixels
-    max_shifts = [int(a/b) for a, b in zip(max_shift_um, dxy)]
+    #max_shifts = [int(a/b) for a, b in zip(max_shift_um, dxy)]
+    max_shifts = (20,20)
     # start a new patch for pw-rigid motion correction every x pixels
-    strides = tuple([int(a/b) for a, b in zip(patch_motion_um, dxy)])
+    #strides = tuple([int(a/b) for a, b in zip(patch_motion_um, dxy)])
+    strides = (32,32)
     # overlap between pathes (size of patch in pixels: strides+overlaps)
-    overlaps = (24, 24)
+    overlaps = (18, 18)
     # maximum deviation allowed for patch with respect to rigid shifts
-    max_deviation_rigid = 3
+    max_deviation_rigid = 5
+
+
+  # estimate minimum of movie
+    min_mov = np.min(tif.imread(fname[0]))
 
     mc_dict = {
         'fnames': fnames,
@@ -123,7 +156,8 @@ def main():
         'strides': strides,
         'overlaps': overlaps,
         'max_deviation_rigid': max_deviation_rigid,
-        'border_nan': 'copy'
+        'border_nan': 'min',
+        'min_mov': min_mov
     }
 
     opts = params.CNMFParams(params_dict=mc_dict)
@@ -133,64 +167,54 @@ def main():
     c, dview, n_processes = cm.cluster.setup_cluster(
         backend='local', n_processes=n_processes, single_thread=False)
 
-    print('checkpoint 1: mcorrect start')
+    print('checkpoint 1: mcorrect start', flush=True)
+
 
     # %%% MOTION CORRECTION
     # first we create a motion correction object with the specified parameters
-    mc = MotionCorrect(fname[0], dview=dview, **opts.get_group('motion'))
+    mc = MotionCorrect(fname, dview=dview, **opts.get_group('motion'))
     # note that the file is not loaded in memory
 
     # %% Run (piecewise-rigid motion) correction using NoRMCorre
-    mc.motion_correct(save_movie=True)
-
-    print('checkpoint 2: mcorrect finish; temp file stored in {}'.format(mc.mmap_file[0]))
-    print('Save corrected tif stack to {}'.format(outfile))
-    
-    (mov, frame_dims, num_frames) = cm.load_memmap(mc.mmap_file[0])
-    mov = mov.reshape( frame_dims + (num_frames,), order='C')
-    # mov is currently in (height, width, nframes format)
-
-
-    # currently format output based on input (stack or single-file)
-    if in_file_ext in ['.tif', '.tiff']:
-        tif.imsave(outfile, mov, imagej=True)
-    else:
-        if not os.path.exists(outfile):
-            os.mkdir(outfile)
-        mov = mov.transpose((2,1,0)) #mov now has nframes in the first dimension
-        for i in range(len(mov)):
-            frame = mov[i]
-            frame_file = os.path.join(outfile, 'frame{:06d}.tiff'.format(i))
-            tif.imsave(frame_file, frame, imagej=True)
-    
-    print('Save complete. Removing temporary files.')
-
-    os.remove(tmpMovPath)
-    os.remove(mc.mmap_file[0])
-
-
-    print('Motion correction script complete.')
-
-    exit()
-    
-    # %% MEMORY MAPPING
+    mc.motion_correct(save_movie=True, template=mc_temp)
     border_to_0 = 0 if mc.border_nan is 'copy' else mc.border_to_0
-    # you can include the boundaries of the FOV if you used the 'copy' option
-    # during motion correction, although be careful about the components near
-    # the boundaries
+ 
+    if pw_rigid:
+        mmap_files = mc.fname_tot_els
+    else:
+        mmap_files = mc.fname_tot_rig
 
-    # memory map the file in order 'C'
-    fname_new = cm.save_memmap(mc.mmap_file, base_name='memmap_', order='C',
-                                border_to_0=border_to_0)  # exclude borders
+    print('mmap file list: {}'.format(mmap_files), flush=True)
 
-    # now load the file
-    Yr, dims, T = cm.load_memmap(fname_new)
-    images = np.reshape(Yr.T, [T] + list(dims), order='F')
-    # load frames in python format (T x X x Y)
+    for n,mmap_file in enumerate(mmap_files):
+        fname_new = cm.save_memmap([mmap_file], base_name='memmap_{}_'.format(n), order='C',border_to_0=border_to_0)
+    
+        (mov, frame_dims, num_frames) = cm.load_memmap(fname_new)
+        mov = mov.reshape( list(frame_dims) + [num_frames]).transpose((2,1,0))
+        mov = mov[channelToKeep::nChannels]
 
-    # %% restart cluster to clean up memory
-    cm.stop_server(dview=dview)
-   
+
+        # currently format output based on input (stack or single-file)
+        #if in_file_ext[n] in ['.tif', '.tiff']:
+        if '.tif' in outfile[n]:
+            tif.imsave(outfile[n], mov, imagej=True)
+        else:
+            print("saving output {}".format(outfile[n]), flush=True)
+            if not os.path.exists(outfile[n]):
+                os.mkdir(outfile[n])
+            for i in range(len(mov)):
+                frame = mov[i]
+                frame_file = os.path.join(outfile[n], 'frame{:06d}.tiff'.format(i))
+                tif.imsave(frame_file, frame, imagej=True)
+    
+    print('Save complete. Removing temporary files.', flush=True)
+
+    if tmpMovPath is not None:
+        os.remove(tmpMovPath)
+    [os.remove(x) for x in mmap_files];
+
+
+    print('Motion correction script complete.', flush=True)
